backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.1
asyncpg==0.30.0
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
PyJWT[crypto]==2.10.1
python-dotenv==1.0.1
spacy==3.8.7
